% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createDataSet}
\alias{createDataSet}
\title{Create a reshaped dataset from the list obtained from
SDMXWrappers:::readXlsxSheets call.}
\usage{
createDataSet(sheetList, colVar, id, time)
}
\arguments{
\item{sheetList}{List object obtained from SDMXWrappers:::readXlsxSheets call}

\item{colVar}{Name of variable that should be used as a column in a reshaped dataset.}

\item{id}{ID variable, typically ISO3 country code}

\item{time}{time variable, typically date.}
}
\value{
data.table with additional attributes containing labeling information.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

