% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readXlsxSheets}
\alias{readXlsxSheets}
\title{Read XLSX file that results from the call to
SDMXWrappers:::sdmxGetAllData function}
\usage{
readXlsxSheets(file)
}
\arguments{
\item{file}{Path to the xlsx file}
}
\value{
List of data.tables. Each list entry contains data from a specific
excel sheet
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

