% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sdmxCollectStaticData}
\alias{sdmxCollectStaticData}
\title{Collect Static Information from RJSDMX:::getSDMX call}
\usage{
sdmxCollectStaticData(o)
}
\arguments{
\item{o}{Object returned by the RJSDMX:::getSDMX function}
}
\value{
data.table with static information
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

