% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sdmxCollectTSData}
\alias{sdmxCollectTSData}
\title{Collect Time Series  from RJSDMX:::getSDMX call}
\usage{
sdmxCollectTSData(o)
}
\arguments{
\item{o}{Object returned by the RJSDMX:::getSDMX function}
}
\value{
data.table with time series.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

