% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sdmxConceptLookup}
\alias{sdmxConceptLookup}
\title{Create a table of concepts within a particular data flow.}
\usage{
sdmxConceptLookup(provider = "IMF", flow = "PGI")
}
\arguments{
\item{provider}{}

\item{flow}{}
}
\value{
List of data.tables contining lookups for each dimension within a
data flow.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

