% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sdmxGetAllFlows}
\alias{sdmxGetAllFlows}
\title{DOWNLOAD AND SAVE ALL DATA FLOWS FROM A GIVEN PROVIDER}
\usage{
sdmxGetAllFlows(provider = "ECB", folder = "/Users/jankocizel/Downloads")
}
\arguments{
\item{provider}{}

\item{folder}{}
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

