% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sdmxLookupDim}
\alias{sdmxLookupDim}
\title{Lookup table of concepts within a particular dimension.}
\usage{
sdmxLookupDim(provider = "IMF", flow = "PGI",
  concept.dimension = "PGI_CONCEPT")
}
\arguments{
\item{provider}{}

\item{flow}{}

\item{concept.dimension}{}
}
\value{
data.table with lookup information
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

